api/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-bulma==0.6.0
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-easy-pdf==0.1.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
heroku==0.1.4
html5lib==1.1
httplib2==0.19.1
idna==2.10
inflect==5.3.0
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
num2words==0.5.10
oauthlib==3.1.0
pdfkit==0.6.1
Pillow==8.2.0
pipenv==2020.11.15
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==1.5
python3-openid==3.2.0
pytz==2020.4
reportlab==3.5.67
requests==2.25.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
whitenoise==5.2.0
wkhtmltopdf==0.2
wrapt==1.12.1
xhtml2pdf==0.2.5
XlsxWriter==1.4.3
